[NNAdapter] Refine the framework of op converter of mediatek_apu, imagination_nna, huawei_kirin_npu and huawei_ascend_npu HALs test=develop
