+++ lite/backends/nnadapter/nnadapter/driver/amlogic_npu/converter/conv2d_transpose.cc
@@ -20,6 +20,7 @@ namespace nnadapter {
 namespace amlogic_npu {
 int ConvertConv2DTranspose(Converter* converter, hal::Operation* operation) {
   CONV_2D_TRANSPOSE_OPERATION_EXTRACT_INPUTS_OUTPUTS
+
   // Convert to amlnpu tensors and operators
   auto input_tensor = converter->GetMappedTensor(input_operand);
   if (!input_tensor) {
