+++ lite/backends/nnadapter/nnadapter/driver/amlogic_npu/converter/converter.cc
@@ -30,7 +30,7 @@ namespace amlogic_npu {
 #define REGISTER_CONVERTER(__op_type__, __func_name__) \
   extern int __func_name__(Converter* converter, hal::Operation* operation);
 #include "driver/amlogic_npu/converter/all.h"  // NOLINT
-#undef __NNADAPTER_DRIVER_AMLOGIC_CONVERTER_ALL_H__
+#undef __NNADAPTER_DRIVER_AMLOGIC_NPU_CONVERTER_ALL_H__
 #undef REGISTER_CONVERTER
 
 int Converter::Apply(hal::Model* model) {
@@ -46,7 +46,7 @@ int Converter::Apply(hal::Model* model) {
     __func_name__(this, operation);                    \
     break;
 #include "driver/amlogic_npu/converter/all.h"  // NOLINT
-#undef __NNADAPTER_DRIVER_AMLOGIC_CONVERTER_ALL_H__
+#undef __NNADAPTER_DRIVER_AMLOGIC_NPU_CONVERTER_ALL_H__
 #undef REGISTER_CONVERTER
       default:
         NNADAPTER_LOG(FATAL) << "Unsupported operation("
