+++ lite/backends/nnadapter/nnadapter/driver/amlogic_npu/converter/pool2d.cc
@@ -20,6 +20,7 @@ namespace nnadapter {
 namespace amlogic_npu {
 int ConvertPool2D(Converter* converter, hal::Operation* operation) {
   POOL_2D_OPERATION_EXTRACT_INPUTS_OUTPUTS
+
   // Convert to amlnpu tensors and operators
   auto input_tensor = converter->GetMappedTensor(input_operand);
   if (!input_tensor) {
