+++ lite/backends/nnadapter/nnadapter/driver/imagination_nna/CMakeLists.txt
@@ -21,7 +21,7 @@ add_definitions(-DNNADAPTER_DRIVER_NAME=${DRIVER_NAME})
 include(dependencies.cmake)
 
 aux_source_directory(converter CONVERTERS)
-set(DRIVER_SRCS imgdnn_manager.cc utility.cc ${CONVERTERS} converter.cc driver.cc)
+set(DRIVER_SRCS imgdnn_manager.cc utility.cc ${CONVERTERS} engine.cc driver.cc)
 set(DRIVER_DEPS ${NNADAPTER_CORE} ${NNADAPTER_UTILITIES} nnadapter_optimizer_symm2asymm ${${DEVICE_NAME}_deps})
 
 add_library(${DRIVER_NAME} SHARED ${DRIVER_SRCS})
