+++ lite/backends/nnadapter/nnadapter/driver/imagination_nna/converter/all.h
@@ -12,24 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef __NNADAPTER_DRIVER_ROCKCHIP_NPU_CONVERTER_ALL_H__  // NOLINT
-#define __NNADAPTER_DRIVER_ROCKCHIP_NPU_CONVERTER_ALL_H__
+#ifndef __NNADAPTER_DRIVER_IMAGINATION_NNA_CONVERTER_ALL_H__  // NOLINT
+#define __NNADAPTER_DRIVER_IMAGINATION_NNA_CONVERTER_ALL_H__
 
 REGISTER_CONVERTER(CONV_2D, ConvertConv2D)
-REGISTER_CONVERTER(CONCAT, ConvertConcat)
 REGISTER_CONVERTER(FULLY_CONNECTED, ConvertFullyConnected)
 REGISTER_CONVERTER(AVERAGE_POOL_2D, ConvertPool2D)
 REGISTER_CONVERTER(MAX_POOL_2D, ConvertPool2D)
-REGISTER_CONVERTER(ADD, ConvertElementwise)
-REGISTER_CONVERTER(SUB, ConvertElementwise)
-REGISTER_CONVERTER(MUL, ConvertElementwise)
-REGISTER_CONVERTER(DIV, ConvertElementwise)
 REGISTER_CONVERTER(SOFTMAX, ConvertSoftmax)
-REGISTER_CONVERTER(SIGMOID, ConvertUnaryActivations)
 REGISTER_CONVERTER(RELU, ConvertUnaryActivations)
 REGISTER_CONVERTER(RELU6, ConvertUnaryActivations)
-REGISTER_CONVERTER(TANH, ConvertUnaryActivations)
-REGISTER_CONVERTER(RESHAPE, ConvertReshape)
-REGISTER_CONVERTER(TRANSPOSE, ConvertTranspose)
 
 #endif  // NOLINT
