+++ lite/backends/nnadapter/nnadapter/driver/imagination_nna/converter/converter.h
@@ -25,36 +25,18 @@
 namespace nnadapter {
 namespace imagination_nna {
 
-class Device {
+class Converter {
  public:
-  Device() {}
-  ~Device() {}
-};
+  explicit Converter(
+      ImgdnnManager* imgdnn_mgr,
+      std::map<hal::Operand*, std::vector<imgdnn_tensor>>* tensors)
+      : imgdnn_mgr_(imgdnn_mgr), tensors_(tensors) {}
+  ~Converter() {}
 
-class Context {
- public:
-  explicit Context(void* device, const char* properties);
-  ~Context();
-
- private:
-  void* device_{nullptr};
-  void* context_{nullptr};
-};
-
-class Program {
- public:
-  explicit Program(Context* context) : context_(context) {}
-  ~Program();
-
-  int Build(hal::Model* model, hal::Cache* cache);
-  int Execute(uint32_t input_count,
-              hal::Argument* input_arguments,
-              uint32_t output_count,
-              hal::Argument* output_arguments);
-
- private:
-  void Clear();
-  // Operand converters
+  // Convert a NNAdapter model to imgdnn graph and tensors
+  int Apply(hal::Model* model);
+  // Mapping a imgdnn tensor to a NNAdapter operand
+  ImgdnnManager* GetImgdnnMgr() { return imgdnn_mgr_; }
   imgdnn_tensor GetMappedTensor(hal::Operand* operand);
   imgdnn_tensor UpdateTensorMap(hal::Operand* operand, imgdnn_tensor tensor);
   imgdnn_tensor AddTensor(int32_t* dimensions,
@@ -89,25 +71,13 @@ class Program {
   imgdnn_tensor ConvertOperand(hal::Operand* operand,
                                std::vector<int32_t> dimensions = {});
 
-  // Operation converters
-  int ConvertConv2D(hal::Operation* operation);
-  int ConvertFullyConnected(hal::Operation* operation);
-  int ConvertPool2D(hal::Operation* operation);
-  int ConvertSoftmax(hal::Operation* operation);
-  int ConvertActivation(hal::Operation* operation);
-
  private:
-  Context* context_{nullptr};
-  // Map NNAdapter operand to imgdnn tensor
-  std::map<hal::Operand*, std::vector<imgdnn_tensor>> tensors_;
-  std::vector<NNAdapterOperandType> input_types_;
-  std::vector<NNAdapterOperandType> output_types_;
-  std::vector<imgdnn_input> input_info_;
-  std::vector<imgdnn_output> output_info_;
-  std::vector<std::pair<imgdnn_memory, size_t>> input_memory_;
-  std::vector<std::pair<imgdnn_memory, size_t>> output_memory_;
-  ImgdnnManager imgdnn_mgr_;
+  ImgdnnManager* imgdnn_mgr_{nullptr};
+  std::map<hal::Operand*, std::vector<imgdnn_tensor>>* tensors_;
 };
+
+#define ADD_OPERATOR(__operator__, ...) \
+  converter->GetImgdnnMgr()->__operator__(__VA_ARGS__)
 
 }  // namespace imagination_nna
 }  // namespace nnadapter
