+++ lite/backends/nnadapter/nnadapter/driver/imagination_nna/converter/unary_activations.cc
@@ -12,45 +12,35 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "driver/imagination_nna/converter.h"
+#include "core/operation/unary_activations.h"
+#include "driver/imagination_nna/converter/converter.h"
 #include "utility/debug.h"
 #include "utility/logging.h"
 
 namespace nnadapter {
 namespace imagination_nna {
 
-int Program::ConvertActivation(hal::Operation* operation) {
-  auto& input_operands = operation->input_operands;
-  auto& output_operands = operation->output_operands;
-  auto input_count = input_operands.size();
-  auto output_count = output_operands.size();
-  NNADAPTER_CHECK_EQ(input_count, 1);
-  NNADAPTER_CHECK_EQ(output_count, 1);
-  // Input
-  auto input_operand = input_operands[0];
-  NNADAPTER_VLOG(5) << "input: " << OperandToString(input_operand);
-  // Output
-  auto output_operand = output_operands[0];
-  NNADAPTER_VLOG(5) << "output: " << OperandToString(output_operand);
+int ConvertUnaryActivations(Converter* converter, hal::Operation* operation) {
+  UNARY_ACTIVATIONS_OPERATION_EXTRACT_INPUTS_OUTPUTS
 
   // Convert to imgdnn tensors and operators
-  auto input_tensor = GetMappedTensor(input_operand);
+  auto input_tensor = converter->GetMappedTensor(input_operand);
   if (!input_tensor) {
-    input_tensor = ConvertOperand(input_operand);
+    input_tensor = converter->ConvertOperand(input_operand);
   }
   imgdnn_tensor output_tensor;
   if (operation->type == NNADAPTER_RELU) {
-    output_tensor =
-        imgdnn_mgr_.CreateReLULayer(input_tensor, true, 0.0, false, 0.0, false);
+    output_tensor = ADD_OPERATOR(
+        CreateReLULayer, input_tensor, true, 0.0, false, 0.0, false);
   } else if (operation->type == NNADAPTER_RELU6) {
-    output_tensor =
-        imgdnn_mgr_.CreateReLULayer(input_tensor, true, 0.0, true, 6.0, false);
+    output_tensor = ADD_OPERATOR(
+        CreateReLULayer, input_tensor, true, 0.0, true, 6.0, false);
   } else {
     NNADAPTER_LOG(FATAL) << "Unsupported activation unary operation type "
                          << OperationTypeToString(operation->type)
                          << " is found.";
   }
-  UpdateTensorMap(output_operand, output_tensor);
+  converter->UpdateTensorMap(output_operand, output_tensor);
   return NNADAPTER_NO_ERROR;
 }
