+++ lite/backends/nnadapter/nnadapter/driver/imagination_nna/engine.h
@@ -1,0 +1,72 @@
+// Copyright (c) 2019 PaddlePaddle Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <map>
+#include <memory>
+#include <string>
+#include <utility>
+#include <vector>
+#include "driver/imagination_nna/imgdnn_manager.h"
+#include "driver/imagination_nna/utility.h"
+
+namespace nnadapter {
+namespace imagination_nna {
+
+class Device {
+ public:
+  Device() {}
+  ~Device() {}
+};
+
+class Context {
+ public:
+  explicit Context(void* device, const char* properties);
+  ~Context();
+
+ private:
+  void* device_{nullptr};
+  void* context_{nullptr};
+};
+
+class Program {
+ public:
+  explicit Program(Context* context) : context_(context) {}
+  ~Program();
+
+  int Build(hal::Model* model, hal::Cache* cache);
+  int Execute(uint32_t input_count,
+              hal::Argument* input_arguments,
+              uint32_t output_count,
+              hal::Argument* output_arguments);
+
+ private:
+  void Clear();
+
+ private:
+  Context* context_{nullptr};
+  // Map NNAdapter operand to imgdnn tensor
+  std::map<hal::Operand*, std::vector<imgdnn_tensor>> tensors_;
+  std::shared_ptr<ImgdnnManager> imgdnn_mgr_{nullptr};
+  std::vector<NNAdapterOperandType> input_types_;
+  std::vector<NNAdapterOperandType> output_types_;
+  std::vector<imgdnn_input> input_info_;
+  std::vector<imgdnn_output> output_info_;
+  std::vector<std::pair<imgdnn_memory, size_t>> input_memory_;
+  std::vector<std::pair<imgdnn_memory, size_t>> output_memory_;
+};
+
+}  // namespace imagination_nna
+}  // namespace nnadapter
