+++ lite/backends/nnadapter/nnadapter/driver/rockchip_npu/CMakeLists.txt
@@ -20,7 +20,7 @@ add_definitions(-DNNADAPTER_DRIVER_NAME=${DRIVER_NAME})
 
 include(dependencies.cmake)
 
-aux_source_directory(./converter CONVERTERS)
+aux_source_directory(converter CONVERTERS)
 aux_source_directory(optimizer OPTIMIZERS)
 set(DRIVER_SRCS utility.cc ${OPTIMIZERS} ${CONVERTERS} engine.cc driver.cc)
 set(DRIVER_DEPS ${NNADAPTER_CORE} ${NNADAPTER_UTILITIES} nnadapter_optimizer_symm2asymm ${${DEVICE_NAME}_deps})
