+++ lite/backends/nnadapter/nnadapter/driver/rockchip_npu/converter/converter.cc
@@ -31,7 +31,7 @@ namespace rockchip_npu {
 #define REGISTER_CONVERTER(__op_type__, __func_name__) \
   extern int __func_name__(Converter* converter, hal::Operation* operation);
 #include "driver/rockchip_npu/converter/all.h"  // NOLINT
-#undef __NNADAPTER_DRIVER_ROCKCHIP_CONVERTER_ALL_H__
+#undef __NNADAPTER_DRIVER_ROCKCHIP_NPU_CONVERTER_ALL_H__
 #undef REGISTER_CONVERTER
 
 int Converter::Apply(hal::Model* model) {
@@ -47,7 +47,7 @@ int Converter::Apply(hal::Model* model) {
     __func_name__(this, operation);                    \
     break;
 #include "driver/rockchip_npu/converter/all.h"  // NOLINT
-#undef __NNADAPTER_DRIVER_ROCKCHIP_CONVERTER_ALL_H__
+#undef __NNADAPTER_DRIVER_ROCKCHIP_NPU_CONVERTER_ALL_H__
 #undef REGISTER_CONVERTER
       default:
         NNADAPTER_LOG(FATAL) << "Unsupported operation("
